requirements.txt
Django==3.2.6
dj-database-url==0.5.0
django-debug-toolbar==3.2.1
django-hosts==4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
gunicorn==20.1.0
oauth2client==4.1.3
oauthlib==3.1.1
psycopg2==2.9.1
python-dateutil==2.8.2
social-auth-app-django==4.0.0
social-auth-core==4.1.0
whitenoise==5.3.0